Normalizer: make delta_namespace work over full names
before this patch, using `delta_namespace ["Module"]` would unfold
every fv whose printing began with "Module". This means that definitions
in "ModuleX", also get expanded, which is not what we expect.

This patch makes it so that it unfold everything in "Module" and it
submodules ("Module.A", "Module.A.B.C", ...) but not other modules
that happen to begin with "Module".

Fixes #3090