book: Typeclasses.fst: remove extra arg!
Now causes an error, the reason is quite surprising though, without the
previous patch the printable_option instance was inferred to have type:

    #a: Type -> #b: printable a -> {| _: printable a |} -> Prims.Tot (printable (option a))

since it was forcing the type of `b` during solving. Then this instance
could be used just fine to solve `printable (option a)`, just at some
extra cost. Now `b` would remain unconstrained and hence unable to be
solved, as it should.