Rel: --MLish implies no SMT allowed
This gives us more type safety in the compiler, since otherwise
_any_ type equalities that could potentially be proven by SMT are
accepted (since we are --lax too). This means for instance that
every polymorphic variable can unify with a concrete type, and the
typechecker would have accepted it.

By running with smt_ok=false, the unifier will simply reject it earlier.