Tc: Rel: in flex-flex quasi, retain meta-tac from the LHS
If we happen to unify the implicits for two typeclass constraints
(likely the same one), we need to retain the tag so this has
any chance of being solved. This changes makes the unifier
retain the meta tag, with a preference for the left one, which
is of course totally arbitrary.