Tc: do not force trivial guard when checking attributes
Trying to force it would make typeclass instantiation run now,
when there are possibly yet-unresolved variables in the environment,
which would prevent the tactic from running. See the Bug3130f example.

This patch makes the typechecker just accumulate the guard, and force it
at the end, where the free uvars restriction will cause the constraints
to be solved in order.