Tc: running meta args in uvar's env, not the original env
We had previously made the choise to run meta argument tactics in the
environment in which they were introduced, instead of the environment of
the uvar that the meta is attached to. These two can differ if the uvar
is unified with other uvars, possibly in more constrained environments.

But, this leads to all sorts of weirdness, such as getting trying
to solve a uvar in an environment that contains itself, if we have
something like:

   let v : ty ?u1 = ... in
   let w : ty ?u2 = ... in
   (* something that unifies the type of v and w *)

We may have set ?u2:=?u1 and copied the meta over, so when we try to
run it to solve ?u1, we should not have ?u1 in scope.