Tc: Allow for #_ to solve instances
When using typeclasses and there is a need to disambiguate one implicit
out of many, say the 3rd, we either have to provide the dictionaries of
the preceding args explicitly:

  f #_ #d_int #bool

or call `solve` like

  f #int #solve #bool

Since just using `#_` for the dictionary will fail, as it will try to be
solved only by unification. This PR makes it so that using `#_` for an
implicit that has a meta (tactic or attr) retains the meta, and is then
solved exactly as if nothing was provided. Then we can disambiguate the
function above as

  f #_ #_ #bool