SMT: Stabilize hint hashes
I recently added some text to the query files to state the versions of
F* and Z3 used, but that includes F*'s commit hash which makes the hint
hashes invalid after any F* commit.

This fixes that problem by using Captions, which get ignored when
printing the query for the purpose of computing the hash.