NBE: mantain shape of primops to resugar them properly
Fixes #3185.

In that issue, a primop was encoded to a Lam node with a set of binders
of unit type (essentially eta-expanding it), which caused a failure in
reading it back. This patch introduces a new shape for an encoded lambda
which is a partially applied primop, so reading it back becomes trivial,
and does not require an eta.