FStar.Tactics.V2: embed Pat_Vars by sealing ppname
This (hopefully) fixes issue #3210. Pattern variables (Pat_Var) were
being embedded as raw strings "", but the unembed requires them to be
sealed. As such, the inspected values weren't always being packed.

I think the NamedView wrapper freshens the ppname or something, and
reseals the result -- which is why NamedView was working OK.