Tc: move missing definitions, skip it when loading from cache
Without this, we would fail when doing something like:

    #set-options "--warn_error +240"
    val f : ...
    [@@expect_failure] let f = ...

and checking it twice. The first time around, the options are processed
and used. The second time, we just load from cache, and then try to find
missing definitions, which gets triggered. But we've already checked the
module, so just skip this check when loading from cache.