Pulse.Lib.HashTable: prevent internal overflows by bounding size
Internally, the hash table performs an addition of two SZ.t, which
are then taken modulo the size of the table. Since the addition can
overflow, the hash table was dynamically checking for an overflow (by
checking that it the addition wouldn't surpass 0xffff, a lower bound for
a max SZ.t) and handling overflows by returning an error code.

That is of course not ideal, especially since said overflows could
really just be impossible (e.g. on a table of size 256).

This patch changes the interface to only allow creating tables with a
size `l` such that `2*l` fits in a SZ.t, which prevents any possible
overflow. With the SizeT module as it is now this means we can always
create tables up to size 0x7fff, and larger ones by assuming `fits_u32`
or `fits_u64`.

This simplifies the code somewhat too. It also removes some impossible
branches from DPE.