Plugin: fix arity mismatch
Normal (Tot) plugins were expecting to be called with a full set
of arguments, including type arguments to be dropped, while
tactic plugins expected to be called on an already trimmed list. The
registry code was matching on the list of arguments and finding the
type arguments, but then passing them as-is to the interpretation
function. This meant that polymorphic Tac plugins were broken.

This fixes it by making all interpretation functions expected a trimmed
list.