Options: prims never extracts
Not having this here causes it to appear in the ALL_ML_FILES (and
others) and trigger an extraction run in most makefiles. But, we then
rule prims out anyway in extraction, so the extraction is attempted
again and again.

This will rule it out from the ALL_ML_FILES and related variables.