Introduce --ext extraction_inline_all
This extension will mark every local definition as CInline for karamel
extraction. Karamel will then consider inlining the letbinding, whenever
this is sound (e.g. it is a pure value) and desirable (e.g. there is no
duplication).

This usually leads to C code that is much more compact. We have been
using it in Kuiper and EverParse with good results.

Related karamel PRs:
- https://github.com/FStarLang/karamel/pull/469
- https://github.com/FStarLang/karamel/pull/470

This F* option behaves similarly to the proposed -faggresive-inlining
option, but it's performed at the F* level instead of karamel.