Options: --already_cached should not consider the current module
If no checked files are present, and we run

    fstar.exe --already_cached '*' A.fst

this will fail, since it could not find a checked file for A itself,
while we (arguably) were stating that all *dependencies* of A should
be checked. This behavior makes it difficult to write Makefiles that
use already cached to make sure we are not inadvertedly checking
dependencies within a given project.

This patch changes the behavior so that, if a file was given in the
command line, it is excluded from the already cached requirement.
This means that `--already_cached '*'` essentially becomes "do not
recursively check any dependencies".