Util: fix mkdir for trailing slashes
When calling F* as in
```
$ fstar.exe --cache_dir a/b/
wrong argument given to option `cache_dir`
```
when `a` does not yet exist, F* is supposed to attempt to create the
directory, but it fails due to the trailing slash. The problem is that
when mkdir raises an ENOENT, we attempt to create parent directories,
but we are computing the parent of 'a/b/' to be 'a/b' (since we naively
removed the last slash component). So the process is:

- Try to mkdir a/b/, get ENOENT
- Try to mkdir a/b, get ENOENT
- Try to mkdir a, works
- Retry mkdir a/b, works
- Retry mkdir a/b/, fails since it already exists.

This fixes the problem ignoring trailing slashes. Moving to a full,
F*-based filename library would be a lot better.