Autoload extension plugins
If we go over a #lang-xyz, and we cannot find an extension xyz in the
internal table, this tries to find a plugin `xyz.cmxs` in the include
path and loading it automatically.

This is extremely convenient for Pulse, where with fstar.include and
this we can now do:

    fstar.exe share/pulse/examples/ArrayTests.fst

and it just works.