Makefile: fix clean rule
The clean-buildfiles subrule could only work after a build, and failed
on a fresh repo. Given we only needed it to decrease the size of docker
images after a build (which we soon won't do) and didn't have automatic
pruning (we do now), this is not that useful.

Just remove it and clean as before.