Tactics: refactor a bit to avoid a cycle
Tactics.MApply.fsti is using typeclasses, which is only working because
of the non-staged build. Conceptually, this is wrong, as any change to
the typeclass resolution/generation will not be in effect when checking
Tactics.MApply. This becomes a concrete problem in the staged build.

This commit refactors MApply to separate its core (mapply0) from the
typeclass interface. This will allow us to make mapply0 a plugin without
any circularity with the typeclass plugins.