tests: rework makefiles
This is part of the new build, but self-contained enough to be merged
before that. It unifies most makefiles in tests/ into a single
mk/test.mk (mk is a new top-level directory for make infra).

It also disconnects examples/ and tests/. The mutually recursive
inclusions are hard to work with.