Tactics: fix V1.Logic
V2 was reworked to have an interface and point to some lemmas in a
separate module to reduce bloat in callers. The same thing was done for
V1.Logic, but the internal lemmas remained and this caused calls to
these tactics to fail (unless the fst happened to be loaded, which is
usually not the case).

Simply removing the internal private lemmas fixes this.