ulib: move FStar.Stubs.Pprint -> FStar.Pprint
The Stubs namespace is for pointing to internal compiler modules, where
Pprint is implemented (FStarC_Pprint.ml). However, it makes sense for
applications to want to use pretty-printing without any relation to the
compiler internals, so this module should just be a normal module in the
library, which is what this patch does.

Ideally, we would also remove the FStarC.Pprint.fsti from src/, and use
this, but there are some problems wrt effect polymorphism. So, we just
duplicate them for now (the interface was already duplicated partly due
to that).