Syntax/Tc: Improving error range for expected failures
When the set of errors do not match the expected set, we will highlight
the @@expect_failure attribute instead of the whole definition.

This works by adding a range to the Sig_fail sigelt, which is modifying
the internal AST. However, since these Sig_fail signatures never make it
to a .checked file, I don't think we need to bump the
cache_version_number.