Remove Cygwin path conversion
F* should not try to do this for the user, and running an external
process for this is definitely not ideal.

Also it's not clear what advantage this had: in Windows, we would not
have a cygpath to call, and in Cygwin, we can properly interpret both
POSIX and Windows paths.