SMTEncoding: prevent hash inconsistencies due to filepaths
The string_of_range call in this function may depend on absolute
filepaths, and be inconsistent across different systems or different
F* install within a single system. Noticed this due to a failing proof
in FStar.Math.Fermat. Looking at the SMT queries for runs of F* in
different directories, the hashes for many Tm_refine_ terms differed.

With this change, there is no difference between the two SMT2 files
(modulo comments) for each run.