Parser.Dep: do not fail if can't find base modules
These calls to module_name_of_file cannot return anything other than
these strings, but can fail if the files cannot be found in the include
path, *even* if we are running with --no_default_includes. This change
makes sure we can work and run dep even if we cannot find a library
(yet?).