FStarC.Util: make sure executable_name and exec_dir point to concrete paths
When building F* from the repo, the bin/fstar.exe symlink does not work
in MacOS due to executable_name being a path with unresolved links, so
trying to locate the library fails. This fixes it by using Unix.realpath
to canonicalize and resolve the path.

Thanks to @danelahman for reporting and testing.