build: simplify, coallescing src/ml/full into ulib/ml/plugin, move src/ml/bare -> src/ml
This separation did not make sense, files in both src/ml/full and
ulib/ml/plugin are there for the support of plugins only, and not needed
for the bare compiler. So we can coallesce these directories into
ulib/ml/plugin, and place all compiler ML files in just src/ml.