Tactics.Print: fix namedv_to_string
Interface interleaving was messing this up: in the interface, we were
claiming to print a `Stubs.Reflection.Types.namedv`, but in the body the
opening of `Tactics.NamedView` gets interleaved, and we were instead
defining a printer for `Tactics.NamedView.namedv`. This would fail to
compile in extracted OCaml code.

This is a reminder that we should remove interleaving.