Tactics: introduce postprocess_type
The attributes postprocess_with and postprocess_for_extraction_with
allow a user to use a custom tactic to transform a definition after
typechecking (to something equivalent); usually the used tactic does
some controlled normalization and unfolding. However, the tactic is only
applied on the body of a letbinding and not the type.

This patch adds an attribute `postprocess_type` to indicate that the
tactic should be applied to the type too.