Main: create odir and cache_dir
This is probably undesirable, but it's the old behavior, and some
Makefiles the karamel makefile rely on it (e.g. when using
RESOURCEMONITOR, as we get a redirection into a non-existing dir, and
runlim will not create the directory).

Just retain this for now, at least in a clear location in the sources.
As far as F* cares, this can be removed.