Marking all projectors with an attribute
This makes it easy to control their normalization behavior or look all
of them up in the environment. They also have qualifiers, injected
by the typechecker, but these are not useful for `delta_attr` nor
`delta_qual`, and are different for each individual definition, so it is
impossible to do a normalization call that e.g. unfolds all projectors.