FStarC.Range: always use basenames
In #3740 we discovered a nasty perf bug caused by trying to resolve
filenames in ranges to actual files, triggering the removal of that
feature and moving that logic to error reporting.

Given that, there is no reason to store full paths in ranges, and we
can just save the basenames of files. This also gets us closer to
reproducible, location-independent checked files, which we currently
lack.

For instance, we now don't have any absolute paths (and privacy leaks!)
in the stage2 checked files.

$ strings stage1/fstarc.checked/* | grep guido | wc
   9709    9709  664844
$ strings stage2/fstarc.checked/* | grep guido | wc
      0       0       0

The --ext fstar:no_absolute_paths option is retained, but only affects
how we print out the ranges, not their internals or what goes into
checked files.