Revert "Makefile: only use dune --release on release builds"
For whatever reason, building the library without the --release flag
makes a dune build call in everparse for the ASN1 parser explode with a
stack overflow:
```
$ dune clean && fstar.exe --ocamlenv dune build
File "dune", lines 3-9, characters 0-113:
3 | (executable
4 |   (name ASN1_Parser)
5 |   (libraries
6 |     fstar.lib
7 |   )
8 |   (flags (:standard -w -8-9-11-26-27-33-39-20))
9 | )
Fatal error: exception Stack overflow
```
This is during compilation of the ASN1_X509.ml module, which is
58k lines long, and the offending function where the crash happens
(dparse_cert) is 28k lines long.

This is probably hitting some limitation of ocamlopt, but also a
reminder that we should try to generate more compact code. For now, just
revert this, and pay some time in development builds.

This reverts commit 1ac7ed0c08f7fea8267db0731d2057a0fe929147.