Introduce FStar.Exception.string_of_exn to the library
This allows user programs (including tactic plugins) to print
exceptions. It is not trivial to make this work in interpreted
mode too (i.e. provide a primop for this) since we do not have a
complete embedding/unembedding for exceptions. There is an e_exn in
FStarC.Tactics.Embeddings, but only handles (properly) a few exceptions.