Normalizer: rework norm requests
This allows the normalizer to recognize normalization requests when the
applications nodes are nested (e.g. `((norm_term [delta]) #t) foo`)
instead of only working for flat n-ary applications.

This is also simpler, we can just look at the stack to find the
first 1/2/3 arguments as needed instead of having to rejig.