SMT: use rlimit and seed as part of query hash
This prevents a situation where we save a hint hash for a proof using
a large rlimit (or particular seed), then decrease it (change it), and
still obtain hash matches for it. The proof is then broken without the
hash, but we won't notice until something else changes, breaking the
hash.