Core: fixup use ranges after cache hit
When we cache the type of a term, we are also caching all the internal
ranges of this type. The def ranges should be fine, they are meant to be
the original location where terms are defined, but the use ranges can
vary.

For example, we have
```
val foo : x:int{ref x} -> int

let x = foo 1 + foo 2
```
We may cache the type of the first `foo` to be `x:int{ref x} -> Type`.
Here, the `ref` inside the refinement has use range equal to the
position of the first foo. When we reuse this cached type for the second
ocurrence, the range will be wrong.

This doesn't quite happen in F* source code as we are not using the
core checker, which is the one that is cached. It does happen in Pulse,
though, which uses reflection typing pritimives that make use of the
core checker.

The patch here simply overrides the use ranges of the cached type for
the use range of the term being checked. This at least guarantees that
the ranges are not pointing to random places

Fixes FStarLang/pulse#416