Parser: allowing refinements on multibinders
This allows to define a function like:

```fstar
let sub (x y : int{x >= y}) : nat =
  x - y
```

where the refinement is only attached to the second binder. This makes
it simple to state relations between several binders of the same type.
This currently fails due to trying to attach the refinement to *all*
binders, which has scoping problems (and would anyway imply repetition).