check-world: remove CBOR job
The cbor branch of Everparse is currently broken as work on has been
progressing on a different branch. Everparse+CBOR is anyway using a
submodule for F*, so a breakage here is not a huge problem, and this
second CBOR branch will be merged into the main Everparse branch soon.

So, disable this job so we can get a green on check-world again.

cc @tahina-pro.