Separate the --warn_error parser, fix crash on ill-formed options
The parsing of the arguments to --warn_error was done via the same
Menhir grammar as normal F*, and the same lexer. This coupling is
needless and slightly confusing. So separate it into a new Menhir parser
and lexer.

Also fix a crash when the string fails to parse, providing a more reasonale
error instead.

This will need a Pulse patch as it uses F* tokens, and RANGE is gone.