Interleave: do not interleave pragmas in IDE mode, just like batch
We've long ignored the pragmas in an fsti when checking an fst
(https://github.com/FStarLang/FStar/pull/1142). This helps in preventing
tons of confusion when checking an fst file, as otherwise the
rlimit/fuel/ifuel can change under the user's feet as they check a file.

However, in the IDE mode, when the pragma was near the top of the
fsti (more specifically in the first chunk that would be interleaved)
in the fst, this was still being interleaved. This is since when try
to find the relevant decls to interleave from the fsti before the
top-level `module Foo` decl, and that returns the empty set, so they
remaing in the `remaining_iface_vals` which are still returned, but were
unfiltered.

This lead to a long headache just now while noticing a file failing in
the IDE, but that had succeeded repeatedly in batch mode.

Fix this by just filtering out all the pragmas before any kind of
interleaving, which should be more robust.