Remove FStarC.BigInt
The F* compiler is an F* program, and `int` is already a bigint. My
recollection is that this module was here for the F# build, which also
needed to use big integers and would not by default. But that is now
gone, so we can get rid of this.

Some operations are being kept in `FStarC.Int.Extra` as they are not in
the (user) library.