Typeclasses: solving refined, unit, and squash goals too
This makes typeclass resolution more aware of refinements in the goals
and in the instances arguments, and tries to handle them.

The logic is roughly to do a pass as before, and if we are about to
fail, we attempt to unrefine the goal and do another pass. Unrefining
means we ignore the refinement formula for the next pass, but whatever
solution we pick for the goal will be checked to satisfy it (via SMT,
later).