ToSyntax: remove generation of projector/discriminator decls
As a follow up to #3954, make desugaring not do anything for
projectors/discriminators except marking that the name is in fact in
scope. This removes a decent chunk of code.

This also makes it such that, once the type declaration sigelt is
pushed to the DsEnv, all the projectors/discriminators will already be
in-scope. This will be important in a follow-up PR.