Use a custom tactic for finding non_informative proofs
Calling tcresolve works, and allows us to use the typeclass system to
extend the set of non-informative types, but comes at a cost. For once,
tcresolve will try to apply local instances from the enviroment beofre
attempting to apply global ones, which means traversing the entire
environment and, for each binder, normalizing it and attempting to unify
it with the goal.

tcresolve should probably have smarter defaults, but, we also don't need
to call it directly. This PR adds a custom tactic that has a fastpath
for the common cases and only calls tcresolve if it could not solve the
goal.