Options: make --already_cached ReverseAccumulated
Similarly to other options, `--already_cached +A --already_cached -A.B`
should require that all of A, except those modules under subnamespace B,
are cached. But this was not the case, the options were interpreted
in backwards order, making this equivalent to `--already_cached -A.B,+A`
(so essentially +A).