SizeT: relax type of fits, prevents spurious queries about non-negativity
When mentioning `SZ.fits x` in a specification, there is a guard
generated for `x` being a nat. This is inconvenient in some settings,
particularly if non-linear arithmetic is involved. For instance, a type
like:
```fstar
let foo (x y z : nat{SZ.fits (x * y * z)} ...
```
requires a proof that a product of three nats is a nat, which can
actually *fail* in contexts where the SMT is too overloaded. (Which is
pretty bad, but an existing problem.)

However, there is no need to restrict `x` to be a nat to make the
predicate well-defined. We can just push the `x >= 0` requirement into
the definition of this predicate, making `SZ.fits (-1)` well-defined
(and false).