Some improvements to typeclass resolution performance
Turns out we were spending a big amount of time normalizing the types
of instances (and local hypotheses) which we were doing since PR
allow typeclass aliases, but way too expensive due to the cost of
normalization. This patch makes this normalization happen only when the
head FV of the return type does not seem to be a class already.

Also, this patch creates a more structured state of global instances,
indexed by the class they belong to. This way the search can just find
the relevant set of instances and try them, instead of going through all
instances (for every class).

It would be very nice to cache this state, which remains valid as long
as no new instances are added, but that's for another day.