Introducing `(#a : Type | deq)` syntax for types + constraints
This syntax simplifies writing several constraints over the same type.
The constraints can just be added to the set after the bar, and many of
them can be given. E.g.
```fstar
let foo5
  (#a #b : Type | deq, ord)
  ...
```
This has two implicit arguments for the types, then `deq` and `ord`
constraints for `a`, and then for `b`. The constraints could also be
partially applied