Rel: do not simplify guards to true for tactics
When running a tactic in phase1 (like any tactic that solves implicits),
this patch would make the preconditions of the lemma become trivial and
disappear as goals, breaking code that expected them to be there. This
is a limitation of the tactics engine, which presents goals as a stack
without a good way of obtaining labels or handles for them. This broke
the Steel tactic as it is contains many applications of lemmas followed
by just the right combination of calls to subtactics or `dismiss`. A
better way of writing robust tactics would be ideal.

Fixes 41ffe463e7834677d70b8b6deab8ee089070b645