language extensions: don't insert #lang-x, use IDE filename when=<input>
Earlier commit added #lang-x to start of parsed module, but this was
unnecessary, so have removed it.

In IDE mode, we might load other modules if the .checked doesn't exist.
The previous commit was always prioritising the IDE filename over the
passed filename, but this was overwriting normal filenames. To resolve
this, I only prioritise IDE filename if the normal filename is
"<input>".