Windows package: override the libgmp DLL PATH when building the package
This DLL must be shipped with the Windows binary package. The current
code is calling `which libgmp-10.dll` to find the DLL in the system, and
chooses this binary to ship. However, in GitHub's Windows runners, this
seems to choose an old version of the DLL that does not work outside of
the Cygwin environment (which we use to build).

This patch tries to find the correct DLL by using lld on the built
executable, within Cygwin, to obtain the DLL.

This makes the Windows workflow work in cloud runners.