Syntax.Util: term_eq: ignore return annotations in match, fix infinite loop in Rel
For two identical matches with a returns annotation, term_eq falls
through into the bottom case and return false. This leads
`head_matches_delta` in the unifier to falsely believe it is making
progress, and go into an infinite loop. We can just ignore these
annotations, similarly to how we ignore ascriptions in this function.

Reported by @santicoronel, in a Pulse example.